Make Capabilities into specific types (#37)
Capabilities are now being decoded into specific types
Device decoding uses Capability types to decode
Groups have associated devices updated whenever devices are updated
Device parsing is a bit smarter in terms of reporting errors
Device and Groups are being tested including new Capability and Device types

UI is still a mess, but this should make it easier to clean that up next. The changes introduces for DeviceView and Lighting are just to get things to type check and run, but I will be refactoring a bunch of that shortly.