fix broken test run index page links (#41)
I didn't take into account the fact that I moved the test index page to a sub-directory (/test-runs/) in the links to individual branch runs. Whoops. Made this a configurable value so I can run it locally and in CI and have it work in both environments.

Also it turns out I didn't actually fix the broken overridden dependencies in #40 , so I remedied that for real--I think, I tested it reasonably well this time I feel like! But I thought so last time too apparently